단기통계 AWS S3백업과정 시 이전 절차 실패시 이후 절차로 넘어가지 않도록 임시 처리

+++ jobs/backupShortTermStats.js
@@ -55,10 +55,19 @@ const checkCommand = (command) => {
 
 const moveToAwsS3 = (downloadFilePath) => {
     console.log('================ 4. compress and move to aws s3');
+    console.log('================ 4.1 compress');
     const gzFilePath = downloadFilePath + '.tar.gz';
-    shell.exec('tar -czvf ' + gzFilePath + ' ' + downloadFilePath);
-    shell.exec('aws s3 mv ' + gzFilePath + ' s3://short-term-stats');
-    shell.exec('rm ' + downloadFilePath);
+    const commandCompress = shell.exec('tar -czvf ' + gzFilePath + ' ' + downloadFilePath);
+
+    if(commandCompress.code === 0) {
+        console.log('================ 4.2 move to aws s3');
+        const commandAws = shell.exec('aws s3 mv ' + gzFilePath + ' s3://short-term-stats');
+
+        if(commandAws.code === 0) {
+            console.log('================ 4.3 remove the downloaded file');
+            shell.exec('rm ' + downloadFilePath);
+        }
+    }
 };
 
 module.exports = {backup, renameCommand, downloadCommand, dropCommand};
